added functionality buttons to the qb list
Gave user options to move a qb up, down, or take off list. All buttons work.
If I can add an event listener, I can remove an event listener. 
I'm trying to do that with an if conditional that prevents a user from adding more than 3 qbs. 
I've tried putting it inside it's own function, I've tried tying it to the end of the original event handler, I've tried e.preventDefault(), but still stuck with this part. 
I added in an h2 element with the goal of making the DOM easier to read. It is appending to the DOM but only at end of the body. 
I know why, it's because I'm doing "document.body..", but I've tried adding it to the correct section via class, id, and div. when I do those options I only get error codes. The only element that actually appends it is document.body.